+++ apps/nextjs/src/app/dashboard/games/shared/[id]/[matchId]/_components/scoresheet-table.tsx
@@ -924,7 +924,7 @@ const TotalRow = ({
                     className="text-center"
                     value={firstTeamPlayer.score ?? ""}
                     onChange={(e) => {
-                      const score = parseInt(e.target.value);
+                      const score = parseFloat(e.target.value);
                       const temp = [...players];
                       const tempTeamPlayer = temp.filter(
                         (player) => player.teamId === team.id,
@@ -967,7 +967,7 @@ const TotalRow = ({
                     className="text-center"
                     value={player.score ?? ""}
                     onChange={(e) => {
-                      const score = parseInt(e.target.value);
+                      const score = parseFloat(e.target.value);
                       const temp = [...players];
                       if (temp[index]?.score !== undefined) {
                         temp[index].score = isNaN(score) ? null : score;
@@ -1015,7 +1015,7 @@ const TotalRow = ({
                 className="text-center"
                 value={player.score ?? ""}
                 onChange={(e) => {
-                  const score = parseInt(e.target.value);
+                  const score = parseFloat(e.target.value);
                   const temp = [...players];
                   if (temp[index]?.score !== undefined) {
                     temp[index].score = isNaN(score) ? null : score;
